backend/requirements.txt
openai>=1.0.0
anthropic>=0.3.0
pydantic>=2.11.0
typing-extensions>=4.13.0
python-dotenv>=1.0.0
httpx>=0.23.0
fastapi
uvicorn
requests
